feat: use static list of code languages
