+++ studio/package.json
@@ -24,7 +24,7 @@
     "@deckdeckgo/deck-utils": "^3.4.0",
     "@deckdeckgo/demo": "^2.0.1",
     "@deckdeckgo/drag-resize-rotate": "^2.0.1",
-    "@deckdeckgo/highlight-code": "^2.4.0",
+    "@deckdeckgo/highlight-code": "^2.5.0",
     "@deckdeckgo/inline-editor": "^3.0.2",
     "@deckdeckgo/lazy-img": "^3.0.0",
     "@deckdeckgo/markdown": "^2.0.1",
