+++ studio/src/app/types/editor/prism-language.ts
@@ -1,0 +1,4 @@
+export interface PrismLanguage {
+  language: string;
+  title: string;
+}
