+++ studio/src/assets/assets.json
@@ -123,9 +123,6 @@
   "gif": {
     "exampleSrc": "https://media.giphy.com/media/xUA7baWfTjfHGLZc3e/200w_d.gif"
   },
-  "prism": {
-    "definitionSrc": "https://raw.githubusercontent.com/PrismJS/prism/master/components.json"
-  },
   "chart": {
     "lineSrc": "https://raw.githubusercontent.com/deckgo/deckdeckgo/master/webcomponents/charts/showcase/data-line-chart.csv",
     "lineCompareSrc": "https://raw.githubusercontent.com/deckgo/deckdeckgo/master/webcomponents/charts/showcase/data-line-chart-to-compare.csv",
